+++ contracts/Core/BlockManager.sol
@@ -18,10 +18,10 @@ contract BlockManager is Initializable, ACL, BlockStorage {
     IVoteManager public voteManager;
     IAssetManager public assetManager;
 
-    event BlockConfirmed(uint256 epoch, uint256 stakerId, uint256[] medians, uint256[] ids, uint256 timestamp);
+    event BlockConfirmed(uint32 epoch, uint256 stakerId, uint256[] medians, uint256[] ids, uint256 timestamp);
 
     event Proposed(
-        uint256 epoch,
+        uint32 epoch,
         uint256 stakerId,
         uint256[] ids,
         uint256[] medians,
@@ -30,7 +30,7 @@ contract BlockManager is Initializable, ACL, BlockStorage {
         uint256 timestamp
     );
 
-    modifier checkEpoch(uint256 epoch) {
+    modifier checkEpoch(uint32 epoch) {
         require(epoch == parameters.getEpoch(), "incorrect epoch");
         _;
     }
@@ -65,7 +65,7 @@ contract BlockManager is Initializable, ACL, BlockStorage {
     // stakers elected in higher iterations can also propose hoping that
     // stakers with lower iteration do not propose for some reason
     function propose(
-        uint256 epoch,
+        uint32 epoch,
         uint256[] memory ids,
         uint256[] memory medians,
         uint256 iteration,
@@ -84,10 +84,11 @@ contract BlockManager is Initializable, ACL, BlockStorage {
 
     //anyone can give sorted votes in batches in dispute state
     function giveSorted(
-        uint256 epoch,
+        uint32 epoch,
         uint256 assetId,
         uint256[] memory sorted
     ) external initialized checkEpoch(epoch) checkState(parameters.dispute()) {
+        // require(disputes[epoch][msg.sender].median == 0, "median already found");
         uint256 medianWeight = voteManager.getTotalInfluenceRevealed(epoch, assetId) / (2);
         uint256 accWeight = disputes[epoch][msg.sender].accWeight;
         uint256 lastVisited = disputes[epoch][msg.sender].lastVisited;
@@ -102,6 +103,7 @@ contract BlockManager is Initializable, ACL, BlockStorage {
             accWeight = accWeight + (voteManager.getVoteWeight(epoch, assetId, sorted[i]));
             if (disputes[epoch][msg.sender].median == 0 && accWeight > medianWeight) {
                 disputes[epoch][msg.sender].median = sorted[i];
+                // break;
             }
         }
         disputes[epoch][msg.sender].lastVisited = lastVisited;
@@ -109,20 +111,20 @@ contract BlockManager is Initializable, ACL, BlockStorage {
     }
 
     // //if any mistake made during giveSorted, resetDispute and start again
-    function resetDispute(uint256 epoch) external initialized checkEpoch(epoch) checkState(parameters.dispute()) {
+    function resetDispute(uint32 epoch) external initialized checkEpoch(epoch) checkState(parameters.dispute()) {
         disputes[epoch][msg.sender] = Structs.Dispute(0, 0, 0, 0);
     }
 
-    function getBlock(uint256 epoch) external view returns (Structs.Block memory _block) {
+    function getBlock(uint32 epoch) external view returns (Structs.Block memory _block) {
         return (blocks[epoch]);
     }
 
-    function getBlockMedians(uint256 epoch) external view returns (uint256[] memory _blockMedians) {
+    function getBlockMedians(uint32 epoch) external view returns (uint256[] memory _blockMedians) {
         _blockMedians = blocks[epoch].medians;
         return (_blockMedians);
     }
 
-    function getProposedBlock(uint256 epoch, uint256 proposedBlock)
+    function getProposedBlock(uint32 epoch, uint256 proposedBlock)
         external
         view
         returns (Structs.Block memory _block, uint256[] memory _blockMedians)
@@ -131,18 +133,17 @@ contract BlockManager is Initializable, ACL, BlockStorage {
         return (_block, _block.medians);
     }
 
-    function getProposedBlockMedians(uint256 epoch, uint256 proposedBlock) external view returns (uint256[] memory _blockMedians) {
+    function getProposedBlockMedians(uint32 epoch, uint256 proposedBlock) external view returns (uint256[] memory _blockMedians) {
         _blockMedians = proposedBlocks[epoch][proposedBlock].medians;
         return (_blockMedians);
     }
 
-    function getNumProposedBlocks(uint256 epoch) external view returns (uint256) {
+    function getNumProposedBlocks(uint32 epoch) external view returns (uint256) {
         return (proposedBlocks[epoch].length);
     }
 
     function confirmBlock() public initialized onlyRole(parameters.getBlockConfirmerHash()) {
-        uint256 epoch = parameters.getEpoch();
-
+        uint32 epoch = parameters.getEpoch();
         for (uint8 i = 0; i < proposedBlocks[epoch - 1].length; i++) {
             if (proposedBlocks[epoch - 1][i].valid) {
                 blocks[epoch - 1] = proposedBlocks[epoch - 1][i];
@@ -164,7 +165,7 @@ contract BlockManager is Initializable, ACL, BlockStorage {
     }
 
     function finalizeDispute(
-        uint256 epoch,
+        uint32 epoch,
         uint256 blockId,
         uint256 assetPosInBlock
     ) public initialized checkEpoch(epoch) checkState(parameters.dispute()) {
@@ -202,7 +203,7 @@ contract BlockManager is Initializable, ACL, BlockStorage {
         return true;
     }
 
-    function _insertAppropriately(uint256 epoch, Structs.Block memory _block) internal {
+    function _insertAppropriately(uint32 epoch, Structs.Block memory _block) internal {
         if (proposedBlocks[epoch].length == 0) {
             proposedBlocks[epoch].push(_block);
             return;
