+++ contracts/Core/interface/IBlockManager.sol
@@ -15,7 +15,7 @@ interface IBlockManager {
     // stakers elected in higher iterations can also propose hoping that
     // stakers with lower iteration do not propose for some reason
     function propose(
-        uint256 epoch,
+        uint32 epoch,
         uint256[] memory ids,
         uint256[] memory medians,
         uint256 iteration,
@@ -25,12 +25,12 @@ interface IBlockManager {
     //anyone can give sorted votes in batches in dispute state
 
     function giveSorted(
-        uint256 epoch,
+        uint32 epoch,
         uint256 assetId,
         uint256[] calldata sorted
     ) external;
 
-    function resetDispute(uint256 epoch) external;
+    function resetDispute(uint32 epoch) external;
 
     function isElectedProposer(
         uint256 iteration,
@@ -40,15 +40,15 @@ interface IBlockManager {
 
     function confirmBlock() external;
 
-    function getBlock(uint256 epoch) external view returns (Structs.Block memory _block);
+    function getBlock(uint32 epoch) external view returns (Structs.Block memory _block);
 
-    function getBlockMedians(uint256 epoch) external view returns (uint256[] memory _blockMedians);
+    function getBlockMedians(uint32 epoch) external view returns (uint256[] memory _blockMedians);
 
-    function getLowerCutoffs(uint256 epoch) external view returns (uint256[] memory _lowerCutoffs);
+    function getLowerCutoffs(uint32 epoch) external view returns (uint256[] memory _lowerCutoffs);
 
-    function getHigherCutoffs(uint256 epoch) external view returns (uint256[] memory _higherCutoffs);
+    function getHigherCutoffs(uint32 epoch) external view returns (uint256[] memory _higherCutoffs);
 
-    function getProposedBlockMedians(uint256 epoch, uint256 proposedBlock) external view returns (uint256[] memory _blockMedians);
+    function getProposedBlockMedians(uint32 epoch, uint256 proposedBlock) external view returns (uint256[] memory _blockMedians);
 
-    function getNumProposedBlocks(uint256 epoch) external view returns (uint256);
+    function getNumProposedBlocks(uint32 epoch) external view returns (uint256);
 }
